Added search bar to User view
